add various type changes to avoid problems on systems where int and ssize_t are different sizes.  Make filestruct->lineno a ssize_t (so that we can avoid negative line numbers at the "Go To Line" prompt), current_y a ssize_t (in order to hold the maximum difference between two filestruct->lineno's), totlines a size_t, and change related variables to match
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2782 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/files.c
@@ -511,7 +511,7 @@ void do_insertfile(
     char *ans = mallocstrcpy(NULL, "");
 	/* The last answer the user typed on the statusbar. */
     filestruct *edittop_save = edittop;
-    int current_y_save = current_y;
+    ssize_t current_y_save = current_y;
     bool at_edittop = FALSE;
 	/* Whether we're at the top of the edit window. */
